requirements.txt
ImageHash==4.3.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pytesseract==0.3.10
PyWavelets==1.5.0
scipy==1.11.4
setuptools==68.2.2
wheel==0.42.0
